model fields just get positive integer

+++ algo/home/models.py
@@ -15,10 +15,10 @@ class AlgorithmModel(models.Model):
     option = models.CharField(max_length=1, choices=OPTION_CHOICES)
 
 class DynamicProcessModel(models.Model):
-    process_name = models.IntegerField()
-    arrival_time = models.IntegerField()
-    burst_time = models.IntegerField()
-    priority = models.IntegerField(default=0)
+    process_name = models.PositiveSmallIntegerField()
+    arrival_time = models.PositiveSmallIntegerField()
+    burst_time = models.PositiveSmallIntegerField()
+    priority = models.PositiveIntegerField(default=0)
 
     def __str__(self):
         return f"process with arrival time:{self.arrival_time} and burst time:{self.burst_time}"
